requirements.txt
color-terminal
thesmuggler
typer
coverage
watchdog
tqdm
pytest
coveralls
PyHamcrest
